Move ThreadPoolWorker into worker/multithreading subpackage
